+++ src/AbstractValue.php
@@ -1,6 +1,6 @@
 <?php
 /*
- * Copyright 2021 Cloud Creativity Limited
+ * Copyright 2022 Cloud Creativity Limited
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,6 +14,8 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+
+declare(strict_types=1);
 
 namespace CloudCreativity\Utils\Value;
 
@@ -66,6 +68,21 @@ abstract class AbstractValue implements ValueInterface
     }
 
     /**
+     * Cast the provided value, if it is not null.
+     *
+     * @param mixed|null $value
+     * @return ValueInterface|null
+     */
+    public static function nullable($value): ?ValueInterface
+    {
+        if (is_null($value)) {
+            return null;
+        }
+
+        return static::cast($value);
+    }
+
+    /**
      * AbstractValue constructor.
      *
      * @param mixed $value
@@ -73,10 +90,21 @@ abstract class AbstractValue implements ValueInterface
      */
     public function __construct($value)
     {
-        if (!$this->accept($value)) {
+        if ($this->notAcceptable($value)) {
             throw new ValueException('Expecting a valid value.');
         }
 
         $this->value = $value;
+    }
+
+    /**
+     * Is the value not acceptable?
+     *
+     * @param mixed $value
+     * @return bool
+     */
+    protected function notAcceptable($value): bool
+    {
+        return !$this->accept($value);
     }
 }