+++ tests/Unit/ValueTest.php
@@ -1,6 +1,6 @@
 <?php
 /*
- * Copyright 2021 Cloud Creativity Limited
+ * Copyright 2022 Cloud Creativity Limited
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -15,6 +15,8 @@
  * limitations under the License.
  */
 
+declare(strict_types=1);
+
 namespace CloudCreativity\Utils\Value\Tests\Unit;
 
 use CloudCreativity\Utils\Value\Tests\IntegerValue;
@@ -23,64 +25,68 @@ use CloudCreativity\Utils\Value\ValueException;
 use CloudCreativity\Utils\Value\ValueInterface;
 use PHPUnit\Framework\TestCase;
 
-/**
- * Class ValueTest
- *
- * @package CloudCreativity\Utils\Value
- */
 class ValueTest extends TestCase
 {
 
-    public function testConstruct()
+    public function testConstruct(): void
     {
         $value = new StringValue('abc');
 
         $this->assertSame('abc', (string) $value);
     }
 
-    public function testFluentConstructor()
+    public function testFluentConstructor(): void
     {
         $this->assertTrue(StringValue::create('abc')->is('abc'));
     }
 
-    public function testToString()
+    public function testToString(): void
     {
         $this->assertSame('abc', (string) StringValue::create('abc')->toString());
     }
 
-    public function testCastsSelf()
+    public function testCastsSelf(): void
     {
         $expected = new StringValue('abc');
         $this->assertSame($expected, StringValue::cast($expected));
+        $this->assertSame($expected, StringValue::nullable($expected));
     }
 
-    public function testCastsScalar()
+    public function testCastsScalar(): void
     {
         $expected = new StringValue('abc');
         $this->assertEquals($expected, StringValue::cast('abc'));
+        $this->assertEquals($expected, StringValue::nullable('abc'));
     }
 
-    public function testCastsValueObject()
+    public function testCastsValueObject(): void
     {
         $mock = $this->createMock(ValueInterface::class);
-        $mock->expects($this->once())->method('get')->willReturn('abc');
+        $mock->expects($this->exactly(2))->method('get')->willReturn('abc');
+
         $this->assertEquals(new StringValue('abc'), StringValue::cast($mock));
+        $this->assertEquals(new StringValue('abc'), StringValue::nullable($mock));
     }
 
-    public function testConstructInvalid()
+    public function testNullableWithNull(): void
+    {
+        $this->assertNull(StringValue::nullable(null));
+    }
+
+    public function testConstructInvalid(): void
     {
         $this->expectException(ValueException::class);
         new StringValue('ab');
     }
 
-    public function testIsNotStrict()
+    public function testIsNotStrict(): void
     {
         $value = new StringValue('123');
 
         $this->assertTrue($value->is(123));
     }
 
-    public function testIsAny()
+    public function testIsAny(): void
     {
         $value = new StringValue('abc');
 
@@ -88,20 +94,20 @@ class ValueTest extends TestCase
         $this->assertFalse($value->is('def', 'ab'));
     }
 
-    public function testIsAnyNotStrict()
+    public function testIsAnyNotStrict(): void
     {
         $value = new StringValue('123');
 
         $this->assertTrue($value->is('abc', 123));
     }
 
-    public function testIsNoArguments()
+    public function testIsNoArguments(): void
     {
         $this->expectException(\BadMethodCallException::class);
         StringValue::create('123')->is();
     }
 
-    public function testEmpty()
+    public function testEmpty(): void
     {
         $value = new IntegerValue(0);
 
@@ -109,7 +115,7 @@ class ValueTest extends TestCase
         $this->assertFalse($value->isNotEmpty());
     }
 
-    public function testNotEmpty()
+    public function testNotEmpty(): void
     {
         $value = new IntegerValue(1);